+++ assets/js/general.js
@@ -99,7 +99,6 @@
             }
         });
     }
-
 })(document, jQuery);