+++ config/archive/read-more.php
@@ -43,4 +43,57 @@ return [
 		'label'    => __( 'Show ellipses', 'genesis-customizer' ),
 		'default'  => true,
 	],
+	[
+		'type'     => 'custom',
+		'settings' => 'divider',
+		'default'  => '<hr>',
+	],
+	[
+		'type'     => 'typography',
+		'settings' => 'typography',
+		'label'    => __( 'Typography', 'genesis-customizer' ),
+		'default'  => [
+			'font-family'    => '',
+			'font-size'      => '',
+			'variant'        => '',
+			'line-height'    => '',
+			'letter-spacing' => '',
+			'text-transform' => '',
+		],
+		'output'   => [
+			[
+				'element' => '.more-link',
+			],
+		],
+	],
+	[
+		'type'     => 'custom',
+		'settings' => 'divider',
+		'default'  => '<hr>',
+	],
+	[
+		'type'     => 'multicolor',
+		'settings' => 'colors',
+		'label'    => __( 'Colors', 'genesis-customizer' ),
+		'choices'  => [
+			'link'       => __( 'Link', 'genesis-customizer' ),
+			'link-hover' => __( 'Link Hover', 'genesis-customizer' ),
+		],
+		'default'  => [
+			'link'       => '',
+			'link-hover' => '',
+		],
+		'output'   => [
+			[
+				'choice'   => 'link',
+				'element'  => '.more-link',
+				'property' => 'color',
+			],
+			[
+				'choice'   => 'link-hover',
+				'element'  => '.more-link:hover,.more-link:focus',
+				'property' => 'color',
+			],
+		],
+	],
 ];