+++ config/content/avatar.php
@@ -4,6 +4,29 @@ namespace GenesisCustomizer;
 
 return [
 	[
+		'type'     => 'multicolor',
+		'settings' => 'colors',
+		'label'    => __( 'Colors', 'genesis-customizer' ),
+		'choices'  => [
+			'border' => __( 'Border', 'genesis-customizer' ),
+		],
+		'default'  => [
+			'border' => _get_color( 'white' ),
+		],
+		'output'   => [
+			[
+				'choice'   => 'border',
+				'element'  => '.avatar',
+				'property' => 'border-color',
+			],
+		],
+	],
+	[
+		'type'     => 'custom',
+		'settings' => 'divider',
+		'default'  => '<hr>',
+	],
+	[
 		'type'     => 'slider',
 		'settings' => 'border-radius',
 		'label'    => __( 'Border Radius', 'genesis-customizer' ),
@@ -18,6 +41,30 @@ return [
 				'element'  => '.avatar',
 				'property' => 'border-radius',
 				'units'    => 'px',
+			],
+		],
+	],
+	[
+		'type'     => 'slider',
+		'settings' => 'border-width',
+		'label'    => __( 'Border Width', 'genesis-customizer' ),
+		'default'  => '0',
+		'choices'  => [
+			'min'  => 0,
+			'max'  => 10,
+			'step' => 1,
+		],
+		'output'   => [
+			[
+				'element'  => '.avatar',
+				'property' => 'border-width',
+				'units'    => 'px',
+			],
+			[
+				'element'       => '.avatar',
+				'property'      => 'border-style',
+				'value_pattern' => 'solid',
+				'exclude'       => [ '0' ],
 			],
 		],
 	],