+++ CHANGELOG.md
@@ -1,5 +1,16 @@
 # Genesis Customizer Changelog
 
+## [1.0.2] - 2019-05-19
+* Added: Social menu
+* Added: Search form settings
+* Added: More avatar settings
+* Added: FontAwesome loader
+* Added: Blockquote quotation mark and citation settings
+* Fixed: Featured content more link
+* Fixed: Blog archive post meta not showing
+* Fixed: Gallery styling
+* Fixed: One click demo import font size
+
 ## [1.0.1] - 2019-05-19
 * Remove some starter sites